Addition/rating params (#12)
* update rating post/put, int rate is now float

* lint

* woops left this out of params after a paste

+++ data/migrations/0008_rating.js
@@ -15,7 +15,7 @@ exports.up = async (knex) => {
       .onDelete('cascade')
       .onUpdate('cascade')
       .notNull();
-    table.integer('rate').notNull();
+    table.float('rate').notNull();
     table.string('comment', 180);
   });
 };
